完善客户新增、修改、删除时客户联系时间的同步

+++ src/main/java/cn/klmb/crm/module/system/controller/admin/notify/SysNotifyTemplateController.java
@@ -133,7 +133,7 @@ public class SysNotifyTemplateController {
     @Parameter(name = "id", description = "id", required = true)
     @Operation(summary = "删除任务")
     @PreAuthorize("@ss.hasPermission('system:notify-template:post')")
-    public CommonResult<Boolean> remove(@RequestParam("id") Integer id) {
+    public CommonResult<Boolean> remove(@RequestParam("id") Long id) {
         xxlJobApiUtils.deleteTask(id);
         return success(true);
     }
